+++ types/Batch/BatchCreateParameters.d.ts
@@ -1,0 +1,5 @@
+import { Shipment } from '../Shipment';
+
+export declare interface IBatchCreateParameters {
+  shipments?: Array<Shipment | string>;
+}
