+++ types/Batch/BatchState.d.ts
@@ -1,0 +1,9 @@
+export declare type TBatchState =
+  | 'creating'
+  | 'creation_failed'
+  | 'created'
+  | 'purchasing'
+  | 'purchase_failed'
+  | 'purchased'
+  | 'label_generating'
+  | 'label_generated';
