+++ types/Batch/BatchStatus.d.ts
@@ -1,0 +1,6 @@
+export declare type TBatchStatus =
+  | 'postage_purchased'
+  | 'postage_purchase_failed'
+  | 'queued_for_purchase'
+  | 'creation_failed'
+  | 'created';
