+++ types/Batch/BatchStatuses.d.ts
@@ -1,0 +1,3 @@
+import { TBatchStatus } from './BatchStatus';
+
+export declare type TBatchStatuses = { [key in TBatchStatus]: number };
