+++ types/Batch/index.d.ts
@@ -1,0 +1,6 @@
+export * from './Batch';
+export * from './BatchState';
+export * from './BatchStatus';
+export * from './BatchStatuses';
+export * from './BatchShipment';
+export * from './BatchCreateParameters';
