+++ types/CarrierAccount/CarrierAccountCreateParameters.d.ts
@@ -1,0 +1,5 @@
+import { ParametersToOmitOnCreate } from '../utils';
+import { ICarrierAccount } from './CarrierAccount';
+
+export declare interface ICarrierAccountCreateParameters
+  extends Omit<ICarrierAccount, ParametersToOmitOnCreate> {}
