+++ types/CarrierAccount/CarrierAccountCredentials.d.ts
@@ -1,0 +1,6 @@
+export declare interface ICarrierAccountCredentials {
+  account_number: string;
+  user_id: string;
+  password: string;
+  access_license_number: string;
+}
