+++ types/Customs/CustomsContentsType.d.ts
@@ -1,0 +1,7 @@
+export declare type TCustomsContentsType =
+  | 'documents'
+  | 'gift'
+  | 'merchandise'
+  | 'returned_goods'
+  | 'sample'
+  | 'other';
