+++ types/Customs/CustomsItem/index.d.ts
@@ -1,0 +1,1 @@
+export * from './CustomsItem';
