+++ types/Customs/index.d.ts
@@ -1,0 +1,3 @@
+export * from './CustomsInfo';
+export * from './CustomsItem';
+export * from './CustomsContentsType';
