+++ types/Insurance/InsuranceStatus.d.ts
@@ -1,0 +1,1 @@
+export declare type TInsuranceStatus = 'new' | 'pending' | 'purchased' | 'failed' | 'cancelled';
