+++ types/Insurance/index.d.ts
@@ -1,0 +1,3 @@
+export * from './Insurance';
+export * from './InsuranceStatus';
+export * from './InsuranceCreateParameters';
