+++ types/Pickup/index.d.ts
@@ -1,0 +1,3 @@
+export * from './Pickup';
+export * from './PickupRate';
+export * from './PickupCreateParameters';
