+++ types/Shipment/Options/LabelFormat.d.ts
@@ -1,0 +1,1 @@
+export declare type LabelFormat = 'PNG' | 'PDF' | 'ZPL' | 'EPL2';
