+++ types/Shipment/index.d.ts
@@ -1,1 +1,9 @@
 export * from './ServiceLevel';
+export * from './Rate';
+export * from './Shipment';
+export * from './Options';
+export * from './ShipmentCreateParameters';
+export * from './ShipmentListParameters';
+export * from './PostageLabel';
+export * from './Message';
+export * from './Form';
