+++ types/index.d.ts
@@ -1,295 +1,12 @@
-import { Address, IAddress, IAddressCreateParameters } from './Address';
+import { Address, IAddress } from './Address';
 import { IBaseObject, IDatedObject, IObjectWithId } from './base';
-import { IFee } from './Fee';
-import { IParcel, Parcel } from './Parcel';
-import { ServiceLevel } from './Shipment';
-import { Carrier, ITracker, Tracker } from './Tracker';
-import { DeepPartial, IAllMethodParameters, ParametersToOmitOnCreate } from './utils';
-
-export declare type TInsuranceStatus = 'new' | 'pending' | 'purchased' | 'failed' | 'cancelled';
-
-/**
- * An Insurance object represents insurance for packages purchased both via the EasyPost API as well as shipments purchased through third parties and later registered with EasyPost.
- * An Insurance is created automatically whenever you buy a Shipment through EasyPost and pass insurance options during the Buy call or in a later call to Insure a Shipment.
- *
- * Insurance purchased through the Shipment Buy or Insure endpoints is immediately insured - there is no possibility of rejection based on tracking information, as the package was just created.
- * On the other hand, Insurance purchased on shipments purchased outside of EasyPost requires creation with a tracking code so that EasyPost may confirm the package existence and current shipping status at the time of purchase.
- *
- * Standalone insurance is created in a pending state to help distinguish it from insurance purchased for an EasyPost Shipment.
- * Both kinds of Insurance use the Tracking system to receive periodic updates, and will report those updates to any appropriate Webhooks on file.
- * Standalone insurance will cancel itself if the tracking information for the given tracking code shows evidence of having been shipped anytime before the insurance was purchased.
- *
- * Unlike Shipments within EasyPost, Insurance objects register To and From Address objects according to the destination and ship-from locations of the package.
- * This means that a Shipment with "is_return: true" actually ships to the listed From Address.
- * Insurance does not have a concept of "is_return", so all insurance records refer to their true package destination as "to_address", regardless of whether or not the shipment is a return.
- *
- * @see https://www.easypost.com/docs/api/node#insurance-object
- */
-export declare interface IInsurance extends IObjectWithId<'Insurance'>, IDatedObject {
-  /**
-   * The unique reference for this Insurance, if any
-   */
-  reference: string;
-
-  /**
-   * USD value of insured goods with sub-cent precision
-   */
-  amount: string;
-
-  /**
-   * The insurance provider used by EasyPost
-   */
-  provider: string;
-
-  /**
-   * An identifying number for some insurance providers used by EasyPost
-   */
-  provider_id: string;
-
-  /**
-   * The ID of the Shipment in EasyPost, if postage was purchased via EasyPost
-   */
-  shipment_id: string;
-
-  /**
-   * The tracking code of either the shipment within EasyPost, or provided by you during creation
-   */
-  tracking_code: string;
-
-  /**
-   * The current status of the insurance, possible values are "new", "pending", "purchased", "failed", or "cancelled"
-   */
-  status: TInsuranceStatus;
-
-  /**
-   * The associated Tracker object
-   */
-  tracker: ITracker;
-
-  /**
-   * The associated Address object for destination
-   */
-  to_address: IAddress;
-
-  /**
-   * The associated Address object for origin
-   */
-  from_address: IAddress;
-
-  /**
-   * The associated InsuranceFee object if any
-   */
-  fee: IFee;
-
-  /**
-   * The list of errors encountered during attempted purchase of the insurance
-   */
-  messages: string[];
-}
-
-export declare type TBatchState =
-  | 'creating'
-  | 'creation_failed'
-  | 'created'
-  | 'purchasing'
-  | 'purchase_failed'
-  | 'purchased'
-  | 'label_generating'
-  | 'label_generated';
-
-/**
- * The Batch object allows you to perform operations on multiple Shipments at once.
- * This includes scheduling a Pickup, creating a ScanForm and consolidating labels.
- * Operations performed on Batches are asynchronous and take advantage of our webhook infrastructure.
- *
- * @see https://www.easypost.com/docs/api/node#batch-object
- */
-export declare interface IBatch extends IObjectWithId<'Batch'>, IDatedObject {
-  /**
-   * An optional field that may be used in place of ID in some API endpoints
-   */
-  reference?: string;
-
-  /**
-   * The overall state. Possible values are "creating", "creation_failed", "created", "purchasing", "purchase_failed", "purchased", "label_generating", and "label_generated"
-   */
-  state: TBatchState;
-
-  /**
-   * The number of shipments added
-   */
-  num_shipments: number;
-
-  shipments: IBatchShipment[];
-
-  /**
-   * A map of BatchShipment statuses to the count of BatchShipments with that status. Valid statuses are "postage_purchased", "postage_purchase_failed", "queued_for_purchase", and "creation_failed"
-   */
-  status: TBatchStatuses;
-
-  /**
-   * The label image url
-   */
-  label_url: string;
-
-  /**
-   * The created ScanForm
-   */
-  scan_form: IScanForm;
-
-  /**
-   * The created Pickup
-   */
-  pickup: IPickup;
-}
-
-export declare type TBatchStatus =
-  | 'postage_purchased'
-  | 'postage_purchase_failed'
-  | 'queued_for_purchase'
-  | 'creation_failed'
-  | 'created';
-export declare type TBatchStatuses = { [key in TBatchStatus]: number };
-
-/**
- * @see https://www.easypost.com/docs/api/node#batch-shipment-object
- */
-export declare interface IBatchShipment {
-  /**
-   * The id of the Shipment. Unique, begins with "shp_"
-   */
-  id: string;
-
-  /**
-   * An optional field that may be used in place of ID in some API endpoints
-   */
-  reference?: string;
-
-  /**
-   * The current status. Possible values are "postage_purchased", "postage_purchase_failed", "queued_for_purchase", "creation_failed", and "created"
-   */
-  batch_status: TBatchStatus;
-
-  /**
-   * A human readable message for any errors that occurred during the Batch's life cycle
-   */
-  batch_message: string;
-}
-
-export declare type TCustomsContentsType =
-  | 'documents'
-  | 'gift'
-  | 'merchandise'
-  | 'returned_goods'
-  | 'sample'
-  | 'other';
-
-/**
- * CustomsInfo objects contain CustomsItem objects and all necessary information for the generation of customs forms required for international shipping.
- *
- * Please see the Shipments documentation for examples of including a CustomsInfo object in a shipment.
- *
- * @see https://www.easypost.com/docs/api/node#customs-info-object
- */
-export declare interface ICustomsInfo extends IObjectWithId<'CustomsInfo'>, IDatedObject {
-  /**
-   * "EEL" or "PFC"
-   */
-  eel_pfc: 'EEL' | 'PFC';
-
-  /**
-   * "documents", "gift", "merchandise", "returned_goods", "sample", or "other"
-   */
-  contents_type: TCustomsContentsType;
-
-  /**
-   * (max 255 characters) Human readable description of content.
-   * Required for certain carriers and always required if contents_type is "other"
-   */
-  contents_explanation: string;
-
-  /**
-   * Electronically certify the information provided
-   */
-  customs_certify: boolean;
-
-  /**
-   * Required if customs_certify is true
-   */
-  customs_signer: string;
-
-  /**
-   * "abandon" or "return", defaults to "return"
-   */
-  non_delivery_option: 'abandon' | 'return' | 'return';
-
-  /**
-   * "none", "other", "quarantine", or "sanitary_phytosanitary_inspection"
-   */
-  restriction_type: 'none' | 'other' | 'quarantine' | 'sanitary_phytosanitary_inspection';
-
-  /**
-   * Required if restriction_type is not "none"
-   */
-  restriction_comments: string;
-
-  /**
-   * Describes to products being shipped
-   */
-  customs_items: ICustomsItem[];
-}
-
-/**
- * A CustomsItem object describes goods for international shipment and should be created then included in a CustomsInfo object.
- *
- * @see https://www.easypost.com/docs/api/node#customs-item-object
- */
-export declare interface ICustomsItem extends IObjectWithId<'CustomsItem'>, IDatedObject {
-  /**
-   * Required, description of item being shipped
-   */
-  description: string;
-
-  /**
-   * Required, greater than zero
-   * float
-   */
-  quantity: number;
-
-  /**
-   * Required, greater than zero, total value (unit value * quantity)
-   * float (USD)
-   */
-  value: number;
-
-  /**
-   * Required, greater than zero, total weight (unit weight * quantity)
-   * float (oz)
-   */
-  weight: number;
-
-  /**
-   * Harmonized Tariff Schedule, e.g. "6109.10.0012" for Men's T-shirts
-   *
-   * @see https://hts.usitc.gov/
-   */
-  hs_tariff_number: string;
-
-  /**
-   * SKU/UPC or other product identifier
-   */
-  code: string;
-
-  /**
-   * Required, 2 char country code
-   */
-  origin_country: string;
-
-  /**
-   * 3 char currency code, default USD
-   */
-  currency: string;
-}
+import { Batch } from './Batch';
+import { CarrierAccount } from './CarrierAccount';
+import { Insurance } from './Insurance';
+import { Parcel } from './Parcel';
+import { Pickup } from './Pickup';
+import { IMessage, IRate, IShipment, ServiceLevel, Shipment } from './Shipment';
+import { Carrier, Tracker } from './Tracker';
 
 /**
  * Webhook Events are triggered by changes in objects you've created via the API.
@@ -391,945 +108,6 @@ export declare interface IOrder extends IObjectWithId<'Order'>, IDatedObject {
    * Set true to create as a return
    */
   is_return: boolean;
-}
-
-/**
- * When rating a Shipment or Pickup, some carriers may fail to generate rates.
- * These failures are returned as part of the Shipment or Pickup as part of their messages attribute, and follow a common object structure.
- *
- * It is important to note that the message value for any member of this list comes directly from the carrier, not from EasyPost.
- * This means that if you see an authentication or other non-shipping error here, it is not an issue between you and EasyPost, it is an issue between you and the carrier, or an issue with the given data.
- *
- * @see https://www.easypost.com/docs/api/node#message-object
- */
-export declare interface IMessage {
-  /**
-   * the name of the carrier generating the error, e.g. "UPS"
-   */
-  carrier: string;
-
-  /**
-   * the category of error that occurred. Most frequently "rate_error"
-   */
-  type: string;
-
-  /**
-   * the string from the carrier explaining the problem
-   */
-  message: string;
-
-  /**
-   * the account id of the carrier. Useful if you have multiple accounts with the same carrier
-   */
-  carrier_account_id: string;
-}
-
-/**
- * After a Shipment is successfully created, it will automatically fetch Rates.
- * You can limit the CarrierAccounts to use for rating by passing the carrier_accounts parameter upon Shipment creation.
- *
- * There are three rate types: the actual rate that will be purchased, rate and currency, the published non-discounted rate, list_rate and list_currency, and the rate if purchased from the post office, retail_rate and retail_currency.
- *
- * @see https://www.easypost.com/docs/api/node#rate-object
- */
-export declare interface IRate extends IObjectWithId<'Rate'>, IDatedObject {
-  /**
-   * service level/name
-   * @see https://www.easypost.com/docs/api/node#service-levels
-   */
-  service: string;
-
-  /**
-   * name of carrier
-   */
-  carrier: string;
-
-  /**
-   * ID of the CarrierAccount record used to generate this rate
-   */
-  carrier_account_id: string;
-
-  /**
-   * 	ID of the Shipment this rate belongs to
-   */
-  shipment_id: string;
-
-  /**
-   * the actual rate quote for this service
-   */
-  rate: string;
-
-  /**
-   * currency for the rate
-   */
-  currency: string;
-
-  /**
-   * the retail rate is the in-store rate given with no account
-   */
-  retail_rate: string;
-
-  /**
-   * currency for the retail rate
-   */
-  retail_currency: string;
-
-  /**
-   * the list rate is the non-negotiated rate given for having an account with the carrier
-   */
-  list_rate: string;
-
-  /**
-   * currency for the list rate
-   */
-  list_currency: string;
-
-  /**
-   * delivery days for this service
-   */
-  delivery_days: number;
-
-  /**
-   * date for delivery
-   */
-  delivery_date: string;
-
-  /**
-   * indicates if delivery window is guaranteed (true) or not (false)
-   */
-  delivery_date_guaranteed: boolean;
-}
-
-/**
- * The workhorse of the EasyPost API, a Shipment is made up of a "to" and "from" Address, the Parcel being shipped, and any customs forms required for international deliveries.
- * Once created a Shipment object is used to retrieve shipping Rates and purchase a label.
- *
- * A Shipment created with a valid to_address, from_address, and parcel will automatically populate its rates attribute.
- *
- * @see https://www.easypost.com/docs/api/node#shipment-object
- */
-export declare interface IShipment extends IObjectWithId<'Shipment'>, IDatedObject {
-  /**
-   * An optional field that may be used in place of id in other API endpoints
-   */
-  reference?: string;
-
-  /**
-   * The destination address
-   */
-  to_address: IAddress;
-
-  /**
-   * The origin address
-   */
-  from_address: IAddress;
-
-  /**
-   * The shipper's address, defaults to from_address
-   */
-  return_address: IAddress;
-
-  /**
-   * The buyer's address, defaults to to_address
-   */
-  buyer_address: IAddress;
-
-  /**
-   * The dimensions and weight of the package
-   */
-  parcel: IParcel;
-
-  /**
-   * Information for the processing of customs
-   */
-  customs_info: ICustomsInfo;
-
-  /**
-   * Document created to manifest and scan multiple shipments
-   */
-  scan_form: IScanForm;
-
-  /**
-   * All associated Form objects
-   */
-  forms: IForm[];
-
-  /**
-   * The associated Insurance object
-   */
-  insurance: IInsurance;
-
-  /**
-   * All associated Rate objects
-   */
-  rates: IRate[];
-
-  /**
-   * The specific rate purchased for the shipment, or null if unpurchased or purchased through another mechanism
-   */
-  selected_rate: IRate;
-
-  /**
-   * The associated PostageLabel object
-   */
-  postage_label: IPostageLabel;
-
-  /**
-   * Any carrier errors encountered during rating, discussed in more depth below
-   */
-  messages: IMessage[];
-
-  /**
-   * All of the options passed to the shipment, discussed in more depth below
-   */
-  options: IOptions;
-
-  /**
-   * Set true to create as a return, discussed in more depth below
-   */
-  is_return: boolean;
-
-  /**
-   * If purchased, the tracking code will appear here as well as within the Tracker object
-   */
-  tracking_code: string;
-
-  /**
-   * The USPS zone of the shipment, if purchased with USPS
-   */
-  usps_zone: string;
-
-  /**
-   * The current tracking status of the shipment
-   */
-  status: string;
-
-  /**
-   * The associated Tracker object
-   */
-  tracker: ITracker;
-
-  /**
-   * The associated Fee objects charged to the billing user account
-   */
-  fees: IFee[];
-
-  /**
-   * The current status of the shipment refund process. Possible values are "submitted", "refunded", "rejected".
-   */
-  refund_status: 'submitted' | 'refunded' | 'rejected';
-
-  /**
-   * The ID of the batch that contains this shipment, if any
-   */
-  batch_id: string;
-
-  /**
-   * The current status of the associated BatchShipment
-   */
-  batch_status: TBatchStatus;
-
-  /**
-   * The current message of the associated BatchShipment
-   */
-  batch_message: string;
-}
-
-/**
- *  - FedEx
- *    - (null) - If print_custom_1_code is not provided it defaults to Customer Reference
- *    - PO - Purchase Order Number
- *    - DP - Department Number
- *    - RMA - Return Merchandise Authorization
- *  - UPS
- *    - AJ - Accounts Receivable Customer Account
- *    - AT - Appropriation Number
- *    - BM - Bill of Lading Number
- *    - 9V - Collect on Delivery (COD) Number
- *    - ON - Dealer Order Number
- *    - DP - Department Number
- *    - 3Q - Food and Drug Administration (FDA) Product Code
- *    - IK - Invoice Number
- *    - MK - Manifest Key Number
- *    - MJ - Model Number
- *    - PM - Part Number
- *    - PC - Production Code
- *    - PO - Purchase Order Number
- *    - RQ - Purchase Request Number
- *    - RZ - Return Authorization Number
- *    - SA - Salesperson Number
- *    - SE - Serial Number
- *    - ST - Store Number
- *    - TN - Transaction Reference Number
- *    - EI - Employer's ID Number
- *    - TJ - Federal Taxpayer ID No.
- */
-export declare type TPrintCustomCode =
-  | 'PO'
-  | 'DP'
-  | 'RMA'
-  | 'AJ'
-  | 'AT'
-  | 'BM'
-  | '9V'
-  | 'ON'
-  | '3Q'
-  | 'IK'
-  | 'MK'
-  | 'MJ'
-  | 'PM'
-  | 'PC'
-  | 'RQ'
-  | 'RZ'
-  | 'SA'
-  | 'SE'
-  | 'ST'
-  | 'TN'
-  | 'EI'
-  | 'TJ'; // TODO enum maybe?
-
-export declare type LabelFormat = 'PNG' | 'PDF' | 'ZPL' | 'EPL2';
-
-/**
- * Shipments can have a variety of additional options which you can specify when creating a shipment.
- * The Options object can be populated with the keys below.
- *
- * @see https://www.easypost.com/docs/api/node#options-object
- */
-export declare interface IOptions {
-  /**
-   * Setting this option to true, will add an additional handling charge.
-   * An Additional Handling charge may be applied to the following:
-   *  - Any article that is encased in an outside shipping container made of metal or wood.
-   *  - Any item, such as a barrel, drum, pail or tire, that is not fully encased in a corrugated cardboard shipping container.
-   *  - Any package with the longest side exceeding 60 inches or its second longest side exceeding 30 inches.
-   *  - Any package with an actual weight greater than 70 pounds.
-   */
-  additional_handling: boolean;
-
-  /**
-   * Setting this option to "0", will allow the minimum amount of address information to pass the validation check.
-   * Only for USPS postage.
-   */
-  address_validation_level: string;
-
-  /**
-   * Set this option to true if your shipment contains alcohol.
-   *  - UPS - only supported for US Domestic shipments
-   *  - FedEx - only supported for US Domestic shipments
-   *  - Canada Post - Requires adult signature 19 years or older. If you want adult signature 18 years or older, instead use delivery_confirmation: ADULT_SIGNATURE
-   */
-  alcohol: boolean;
-
-  /**
-   * Setting this option to true will indicate to the carrier to prefer delivery by drone, if the carrier supports drone delivery.
-   */
-  by_drone: boolean;
-
-  /**
-   * Setting this to true will add a charge to reduce carbon emissions.
-   */
-  carbon_neutral: boolean;
-
-  /**
-   * Adding an amount will have the carrier collect the specified amount from the recipient.
-   */
-  cod_amount: string;
-
-  /**
-   * Method for payment. "CASH", "CHECK", "MONEY_ORDER"
-   */
-  cod_method: 'CASH' | 'CHECK' | 'MONEY_ORDER';
-
-  /**
-   * The ID of the Address to which the COD payment should be returned.
-   * Defaults to the origin address.
-   * Only available on FedEx shipments.
-   */
-  cod_address_id: string;
-
-  /**
-   * Which currency this shipment will show for rates if carrier allows.
-   */
-  currency: string;
-
-  /**
-   * If you want to request a signature, you can pass "ADULT_SIGNATURE" or "SIGNATURE".
-   * You may also request "NO_SIGNATURE" to leave the package at the door.
-   *  - All - some options may be limited for international shipments
-   *  - FedEx - "INDIRECT_SIGNATURE" is also an option
-   *  - USPS - additional options
-   *    - "ADULT_SIGNATURE_RESTRICTED"
-   *    - "SIGNATURE_RESTRICTED"
-   */
-  delivery_confirmation:
-    | 'ADULT_SIGNATURE'
-    | 'SIGNATURE'
-    | 'NO_SIGNATURE'
-    | 'INDIRECT_SIGNATURE'
-    | 'ADULT_SIGNATURE_RESTRICTED'
-    | 'SIGNATURE_RESTRICTED';
-
-  /**
-   * Method the customer will use to transfer the package to the carrier.
-   * "REGULAR_PICKUP", "SCHEDULED_PICKUP", "RETAIL_LOCATION", "STATION" or "DROP_BOX".
-   * Supported carriers and their corresponding carrier dropoff codes:
-   *  - FedEx
-   *    - "REGULAR_PICKUP" - "REGULAR_PICKUP" (default)
-   *    - "SCHEDULED_PICKUP" - "REQUEST_COURIER"
-   *    - "RETAIL_LOCATION" - "BUSINESS_SERVICE_CENTER"
-   *    - "STATION" - "STATION"
-   *    - "DROP_BOX" - "DROP_BOX"
-   */
-  dropoff_type:
-    | 'REGULAR_PICKUP'
-    | 'SCHEDULED_PICKUP'
-    | 'RETAIL_LOCATION'
-    | 'STATION'
-    | 'DROP_BOX'
-    | 'REQUEST_COURIER'
-    | 'BUSINESS_SERVICE_CENTER';
-
-  /**
-   * Package contents contain dry ice.
-   *  - UPS - Need dry_ice_weight to be set
-   *  - UPS MailInnovations - Need dry_ice_weight to be set
-   *  - FedEx - Need dry_ice_weight to be set
-   */
-  dry_ice: boolean;
-
-  /**
-   * If the dry ice is for medical use, set this option to true.
-   *  - UPS - Need dry_ice_weight to be set
-   *  - UPS MailInnovations - Need dry_ice_weight to be set
-   */
-  dry_ice_medical: boolean;
-
-  /**
-   * Weight of the dry ice in ounces.
-   *  - UPS - Need dry_ice to be set
-   *  - UPS MailInnovations - Need dry_ice to be set
-   *  - FedEx - Need dry_ice to be set
-   */
-  dry_ice_weight: string;
-
-  /**
-   * Possible values "ADDRESS_SERVICE_REQUESTED", "FORWARDING_SERVICE_REQUESTED", "CHANGE_SERVICE_REQUESTED", "RETURN_SERVICE_REQUESTED", "LEAVE_IF_NO_RESPONSE"
-   */
-  endorsement:
-    | 'ADDRESS_SERVICE_REQUESTED'
-    | 'FORWARDING_SERVICE_REQUESTED'
-    | 'CHANGE_SERVICE_REQUESTED'
-    | 'RETURN_SERVICE_REQUESTED'
-    | 'LEAVE_IF_NO_RESPONSE';
-
-  /**
-   * Additional cost to be added to the invoice of this shipment.
-   * Only applies to UPS currently.
-   */
-  freight_charge: number;
-
-  /**
-   * This is to designate special instructions for the carrier like "Do not drop!".
-   */
-  handling_instructions: string;
-
-  /**
-   * Dangerous goods indicator.
-   * Possible values are "PRIMARY_CONTAINED", "PRIMARY_PACKED", "PRIMARY", "SECONDARY_CONTAINED", "SECONDARY_PACKED", "SECONDARY", "ORMD", "LIMITED_QUANTITY", "LITHIUM".
-   * Applies to USPS, FedEx and DHL eCommerce.
-   */
-  hazmat:
-    | 'PRIMARY_CONTAINED'
-    | 'PRIMARY_PACKED'
-    | 'PRIMARY'
-    | 'SECONDARY_CONTAINED'
-    | 'SECONDARY_PACKED'
-    | 'SECONDARY'
-    | 'ORMD'
-    | 'LIMITED_QUANTITY'
-    | 'LITHIUM';
-
-  /**
-   * Package will wait at carrier facility for pickup.
-   */
-  hold_for_pickup: boolean;
-
-  /**
-   * Incoterm negotiated for shipment.
-   * Supported values are "EXW", "FCA", "CPT", "CIP", "DAT", "DAP", "DDP", "FAS", "FOB", "CFR", and "CIF".
-   * Setting this value to anything other than "DDP" will pass the cost and responsibility of duties on to the recipient of the package(s), as specified by Incoterms rules
-   */
-  incoterm: 'EXW' | 'FCA' | 'CPT' | 'CIP' | 'DAT' | 'DAP' | 'DDP' | 'FAS' | 'FOB' | 'CFR' | 'CIF';
-
-  /**
-   * This will print an invoice number on the postage label.
-   */
-  invoice_number: string;
-
-  /**
-   * Set the date that will appear on the postage label.
-   * Accepts ISO 8601 formatted string including time zone offset.
-   * EasyPost stores all dates as UTC time.
-   */
-  label_date: string;
-
-  /**
-   * Supported label formats include "PNG", "PDF", "ZPL", and "EPL2".
-   * "PNG" is the only format that allows for conversion.
-   *
-   * @see https://www.easypost.com/docs/api#convert-the-label-format-of-a-shipment
-   */
-  label_format: LabelFormat;
-
-  /**
-   * Whether or not the parcel can be processed by the carriers equipment.
-   */
-  machinable: boolean;
-
-  /**
-   * Setting payment type to bill the correct account for purchasing postage.
-   */
-  payment: {
-    /**
-     * Supported values are "SENDER", "THIRD_PARTY", "RECEIVER", "COLLECT". Defaults to SENDER.
-     */
-    type: 'SENDER' | 'THIRD_PARTY' | 'RECEIVER' | 'COLLECT';
-
-    /**
-     * Setting account number.
-     * Required for RECEIVER and THIRD_PARTY.
-     */
-    account: string;
-
-    /**
-     * Setting country code that the account is based in.
-     * Required for THIRD_PARTY.
-     */
-    country: string;
-
-    /**
-     * Setting postal code that the account is based in.
-     * Required for RECEIVER and THIRD_PARTY.
-     */
-    postal_code: string;
-  };
-
-  /**
-   * You can optionally print custom messages on labels.
-   * The locations of these fields show up on different spots on the carrier's labels.
-   */
-  print_custom_1: string;
-
-  /**
-   * An additional message on the label. Same restrictions as print_custom_1
-   */
-  print_custom_2: string;
-
-  /**
-   * An additional message on the label. Same restrictions as print_custom_1
-   */
-  print_custom_3: string;
-
-  /**
-   * Create a barcode for this custom reference if supported by carrier.
-   */
-  print_custom_1_barcode: boolean;
-
-  /**
-   * Create a barcode for this custom reference if supported by carrier.
-   */
-  print_custom_2_barcode: boolean;
-
-  /**
-   * Create a barcode for this custom reference if supported by carrier.
-   */
-  print_custom_3_barcode: boolean;
-
-  /**
-   * Specify the type of print_custom_1.
-   */
-  print_custom_1_code: TPrintCustomCode;
-
-  /**
-   * Specify the type of print_custom_2.
-   */
-  print_custom_2_code: TPrintCustomCode;
-
-  /**
-   * Specify the type of print_custom_3.
-   */
-  print_custom_3_code: TPrintCustomCode;
-
-  /**
-   * Set this value to true for delivery on Saturday.
-   * When setting the saturday_delivery option, you will only get rates for services that are eligible for saturday delivery.
-   * If no services are available for saturday delivery, then you will not be returned any rates.
-   * You may need to create 2 shipments, one with the saturday_delivery option set on one without to get all your eligible rates.
-   */
-  saturday_delivery: boolean;
-
-  /**
-   * This option allows you to request restrictive rates from USPS.
-   * Can set to 'USPS.MEDIAMAIL' or 'USPS.LIBRARYMAIL'.
-   */
-  special_rates_eligibility: 'USPS.MEDIAMAIL' | 'USPS.LIBRARYMAIL';
-
-  /**
-   * You can use this to override the hub ID you have on your account.
-   */
-  smartpost_hub: string;
-
-  /**
-   * The manifest ID is used to group SmartPost packages onto a manifest for each trailer.
-   */
-  smartpost_manifest: string;
-
-  /**
-   * A reference ID for aggregating DHL eCommerce billing data.
-   */
-  billing_ref: string;
-
-  /**
-   * Certified Mail provides the sender with a mailing receipt and, upon request, electronic verification that an article was delivered or that a delivery attempt was made.
-   */
-  certified_mail: boolean;
-
-  /**
-   * Registered Mail is the most secure service that the USPS offers.
-   * It incorporates a system of receipts to monitor the movement of the mail from the point of acceptance to delivery
-   */
-  registered_mail: boolean;
-
-  /**
-   * The value of the package contents
-   */
-  registered_mail_amount: number;
-
-  /**
-   * An electronic return receipt may be purchased at the time of mailing and provides a shipper with evidence of delivery (to whom the mail was delivered and date of delivery), and information about the recipient's actual delivery address.
-   * Only applies to the USPS.
-   */
-  return_receipt: boolean;
-}
-
-/**
- * @see https://www.easypost.com/docs/api/node#form-object
- */
-export declare interface IForm extends IObjectWithId<'Form'>, IDatedObject {
-  /**
-   * The type of form that we returned, can be one of "high_value_report", "commercial_invoice", "cod_return_label", "order_summary", "cn22"
-   */
-  form_type:
-    | 'high_value_report'
-    | 'commercial_invoice'
-    | 'cod_return_label'
-    | 'order_summary'
-    | 'cn22';
-
-  /**
-   * The address we return the form back at
-   */
-  form_url: string;
-
-  /**
-   * If we have submitted the form to the carrier on behalf of the customer
-   */
-  submitted_electronically: boolean;
-}
-
-/**
- * The Pickup object allows you to schedule a pickup from your carrier from your customer's residence or place of business.
- * Supported carriers include:
- *  - Asendia Europe
- *  - Canada Post
- *  - Canpar
- *  - DHL Express
- *  - Endicia
- *  - FedEx
- *  - GSO
- *  - Lasership
- *  - LSO
- *  - Ontrac
- *  - Purolator
- *  - UPS
- *  - USPS
- *
- * After a Pickup is successfully created, it will automatically fetch PickupRates for each CarrierAccount specified that supports scheduled pickups.
- * Then a PickupRate must be selected and purchased before the pickup can be successfully scheduled.
- *
- * @see https://www.easypost.com/docs/api/node#pickup-object
- */
-export declare interface IPickup extends IObjectWithId<'Pickup'>, IDatedObject {
-  /**
-   * An optional field that may be used in place of ID in some API endpoints
-   */
-  reference?: string;
-
-  /**
-   * One of: "unknown", "scheduled", or "canceled"
-   */
-  status: 'unknown' | 'scheduled' | 'canceled';
-
-  /**
-   * The earliest time at which the package is available to pick up
-   */
-  min_datetime: string;
-
-  /**
-   * The latest time at which the package is available to pick up.
-   * Must be later than the min_datetime
-   */
-  max_datetime: string;
-
-  /**
-   * Is the pickup address the account's address?
-   */
-  is_account_address: boolean;
-
-  /**
-   * Additional text to help the driver successfully obtain the package
-   */
-  instructions: string;
-
-  /**
-   * A list of messages containing carrier errors encountered during pickup rate generation
-   */
-  messages: IMessage[];
-
-  /**
-   * The confirmation number for a booked pickup from the carrier
-   */
-  confirmation: string;
-
-  /**
-   * The associated Shipment
-   */
-  shipment: IShipment;
-
-  /**
-   * The associated Address
-   */
-  address: IAddress;
-
-  /**
-   * The list of carriers (if empty, all carriers were used) used to generate pickup rates
-   */
-  carrier_accounts: ICarrierAccount[];
-
-  /**
-   * The list of different pickup rates across valid carrier accounts for the shipment
-   */
-  pickup_rates: IPickupRate[];
-}
-
-/**
- * @see https://www.easypost.com/docs/api/node#pickup-rate-object
- */
-export declare interface IPickupRate extends IObjectWithId<'PickupRate'>, IDatedObject {
-  /**
-   * service name
-   */
-  service: string;
-
-  /**
-   * name of carrier
-   */
-  carrier: string;
-
-  /**
-   * the actual rate quote for this service
-   */
-  rate: string;
-
-  /**
-   * currency for the rate
-   */
-  currency: string;
-
-  /**
-   * the ID of the pickup this is a quote for
-   */
-  pickup_id: string;
-}
-
-/**
- * A CarrierAccount encapsulates your credentials with the carrier.
- * The CarrierAccount object provides CRUD operations for all CarrierAccounts.
- *
- * Each EasyPost account is automatically provided a USPS account managed by EasyPost.
- *
- * Other operations, such as Shipment creation, can reference CarrierAccounts to reduce the scope of data returned.
- * For instance, you may have multiple warehouses that need to use distinct FedEx SmartPost credentials to request the correct rates.
- * Rate objects will include a carrier_account_id field which can be used to determine the account used for rating.
- *
- * @see https://www.easypost.com/docs/api/node#carrier-account-object
- */
-export declare interface ICarrierAccount extends IObjectWithId<'CarrierAccount'>, IDatedObject {
-  /**
-   * The name of the carrier type.
-   */
-  type: string;
-
-  /**
-   * Contains "credentials" and/or "test_credentials", or may be empty
-   */
-  fields: ICarrierAccountFields;
-
-  /**
-   * If clone is true, only the reference and description are possible to update
-   */
-  clone: boolean;
-
-  /**
-   * An optional, user-readable field to help distinguish accounts
-   */
-  description: string;
-
-  /**
-   * An optional field that may be used in place of carrier_account_id in other API endpoints
-   */
-  reference: string;
-
-  /**
-   * The name used when displaying a readable value for the type of the account
-   */
-  readable: string;
-
-  /**
-   * Unlike the "credentials" object contained in "fields", this nullable object contains just raw credential pairs for client library consumption
-   */
-  credentials: ICarrierAccountCredentials;
-
-  /**
-   * Unlike the "test_credentials" object contained in "fields", this nullable object contains just raw test_credential pairs for client library consumption
-   */
-  test_credentials: ICarrierAccountCredentials;
-}
-
-export declare interface ICarrierAccountCredentials {
-  account_number: string;
-  user_id: string;
-  password: string;
-  access_license_number: string;
-}
-
-/**
- * @see https://www.easypost.com/docs/api/node#carrier-account-fields-object
- */
-export declare interface ICarrierAccountFields {
-  /**
-   * Credentials used in the production environment.
-   */
-  credentials: ICarrierAccountField;
-
-  /**
-   * Credentials used in the test environment.
-   */
-  test_credentials: ICarrierAccountField;
-
-  /**
-   * For USPS this designates that no credentials are required.
-   */
-  auto_link: boolean;
-
-  /**
-   * When present, a separate authentication process will be required through the UI to link this account type.
-   */
-  custom_workflow: boolean;
-}
-
-/**
- * @see https://www.easypost.com/docs/api/node#carrier-account-field-object
- */
-export declare interface ICarrierAccountField {
-  /**
-   * Each key in the sub-objects of a CarrierAccount's fields is the name of a settable field
-   */
-  key: string;
-
-  /**
-   * The visibility value is used to control form field types, and is discussed in the CarrierType section
-   */
-  visibility: string;
-
-  /**
-   * The label value is used in form rendering to display a more precise field name
-   */
-  label: string;
-
-  /**
-   * Checkbox fields use "0" and "1" as False and True, all other field types present plaintext, partly-masked, or masked credential data for reference
-   */
-  value: string;
-}
-
-/**
- * A ScanForm can be created to speed up and simplify the carrier pickup process.
- * The ScanForm is one document that can be scanned to mark all included tracking codes as "Accepted for Shipment" by the carrier.
- * The following criteria must met before creation:
- *  - Refunded Shipments cannot be added
- *  - Each Shipment must have the same origin address
- *  - Shipments must all be dated (using the label_date option) on or after the date the form is generated
- *  - Shipments cannot be added to more than one ScanForm
- *  - Existing ScanForms may not be updated with additional Shipments. If a ScanForm already exists, and new Shipments need to be added, a new ScanForm must be created.
- *  - Shipments should be provided in the form of an array
- *
- * @see https://www.easypost.com/docs/api/node#scan-form-object
- */
-export declare interface IScanForm extends IObjectWithId<'ScanForm'>, IDatedObject {
-  /**
-   * Current status.
-   * Possible values are "creating", "created" and "failed"
-   */
-  status: 'creating' | 'created' | 'failed';
-
-  /**
-   * Human readable message explaining any failures
-   */
-  message: string;
-
-  /**
-   * Address that the Shipments will be shipped from
-   */
-  address: IAddress;
-
-  /**
-   * Tracking codes included on the ScanForm
-   */
-  tracking_codes: string[];
-
-  /**
-   * Url of the document
-   */
-  form_url: string;
-
-  /**
-   * File format of the document
-   */
-  form_file_type: string;
-
-  /**
-   * The id of the associated Batch. Unique, starts with "batch_"
-   */
-  batch_id: string;
-}
-
-/**
- * PostageLabel Object
- */
-export declare interface IPostageLabel extends IObjectWithId<'PostageLabel'>, IDatedObject {
-  date_advance: number;
-  integrated_form: string;
-  label_date: string;
-  label_resolution: number;
-  label_size: string;
-  label_type: string;
-  label_url: string;
-  label_file_type: string;
-  label_pdf_url: string;
-  label_epl2_url: string;
-  label_zpl_url: string;
 }
 
 /**
@@ -1552,233 +330,6 @@ export declare interface IUser extends IObjectWithId<'User'> {
   children: IUser[];
 }
 
-/**
- * @see https://www.easypost.com/docs/api/node#retrieve-a-list-of-shipments
- */
-export declare interface IShipmentListParameters extends IAllMethodParameters {
-  /**
-   * Only include Shipments that have been purchased.
-   * Default is true
-   */
-  purchased?: boolean;
-
-  /**
-   * Also include Shipments created by Child Users.
-   * Defaults to false
-   */
-  include_children?: boolean;
-}
-
-export declare interface IShipmentCreateParameters {
-  reference?: string;
-  to_address: DeepPartial<IAddressCreateParameters> | string;
-  from_address: DeepPartial<IAddressCreateParameters> | string;
-  parcel: IParcel | string;
-
-  /**
-   * optional array of ids that begin with "ca_"
-   */
-  carrier_accounts?: string[];
-
-  customs_info?: DeepPartial<ICustomsInfo>;
-  options?: DeepPartial<IOptions>;
-}
-
-export declare class Shipment implements IShipment {
-  public constructor(input: IShipmentCreateParameters);
-
-  reference?: string;
-  to_address: IAddress;
-  from_address: IAddress;
-  return_address: IAddress;
-  buyer_address: IAddress;
-  parcel: IParcel;
-  customs_info: ICustomsInfo;
-  scan_form: IScanForm;
-  forms: IForm[];
-  insurance: IInsurance;
-  rates: IRate[];
-  selected_rate: IRate;
-  postage_label: IPostageLabel;
-  messages: IMessage[];
-  options: IOptions;
-  is_return: boolean;
-  tracking_code: string;
-  usps_zone: string;
-  status: string;
-  tracker: ITracker;
-  fees: IFee[];
-  refund_status: 'submitted' | 'refunded' | 'rejected';
-  batch_id: string;
-  batch_status: TBatchStatus;
-  batch_message: string;
-  id: string;
-  mode: 'test' | 'production';
-  object: 'Shipment';
-  created_at: string;
-  updated_at: string;
-
-  /**
-   * A Shipment is almost exclusively a container for other objects, and thus a Shipment may reuse many of these objects.
-   * Additionally, all the objects contained within a Shipment may be created at the same time.
-   *
-   * The origin/destination Address and Parcel are required for rating.
-   * CustomsInfo is required to rate an international Shipment, this includes when the destination is a US Territory.
-   * The associated Tracker, Fees, and Rates are generated by EasyPost and cannot be modified by the user.
-   *
-   * You can limit the CarrierAccounts to use for rating by passing the carrier_accounts parameter.
-   *
-   * @see https://www.easypost.com/docs/api/node#create-a-shipment
-   */
-  public save(): Promise<Shipment>;
-
-  /**
-   * The Shipment List is a paginated list of all Shipment objects associated with the given API Key.
-   * It accepts a variety of parameters which can be used to modify the scope.
-   * The has_more attribute indicates whether or not additional pages can be requested.
-   * The recommended way of paginating is to use either the before_id or after_id parameter to specify where the next page begins.
-   *
-   * @see https://www.easypost.com/docs/api/node#retrieve-a-list-of-shipments
-   *
-   * @param params
-   */
-  static all(
-    params?: IShipmentListParameters,
-  ): Promise<{ shipments: Shipment[]; has_more: boolean }>;
-
-  /**
-   * A Shipment can be retrieved by either its id or reference.
-   * However it is recommended to use EasyPost's provided identifiers because uniqueness on reference is not enforced.
-   *
-   * @see https://www.easypost.com/docs/api/node#retrieve-a-shipment
-   *
-   * @param shipmentId Unique, begins with "shp_"
-   */
-  static retrieve(shipmentId: string): Promise<Shipment>;
-
-  /**
-   * To purchase a Shipment you only need to specify the Rate to purchase.
-   * This operation populates the tracking_code and postage_label attributes.
-   * The default image format of the associated PostageLabel is PNG.
-   * To change this default see the label_format option.
-   *
-   * Additionally, insurance may be added during the purchase.
-   * To specify an amount to insure, pass the insurance attribute as a string.
-   * The currency of all insurance is USD.
-   *
-   * @see https://www.easypost.com/docs/api/node#buy-a-shipment
-   *
-   * @param rate rate id (begins with "rate_") or rate object
-   * @param insuranceAmount
-   */
-  public buy(rate: string | IRate, insuranceAmount?: number): Promise<Shipment>;
-
-  /**
-   * Refunding a Shipment is available for many carriers supported by EasyPost.
-   * Once the refund has been submitted, refund_status attribute of the Shipment will be populated with one of the possible values: "submitted", "refunded", "rejected".
-   * The most common initial status is "submitted".
-   * Many carriers require that the refund be processed before the refund_status will move to "refunded".
-   * The length of this process depends on the carrier, but no greater than 30 days.
-   *
-   * Refunds created very shortly after a label is generated may be improperly flagged as invalid, but you may retry a refund with the "rejected" status by submitting the same request again.
-   * Carriers that are bill-on-scan tend to have refunds attempts return as "not_applicable", which will not change with multiple retries.
-   *
-   * @see https://www.easypost.com/docs/api/node#refund-a-shipment
-   */
-  public refund(): Promise<Shipment>;
-
-  public lowestRate(carriers?: Carrier[], services?: ServiceLevel[]): IRate;
-
-  /**
-   * A Shipment's PostageLabel can be converted from PNG to other formats.
-   * If the PostageLabel was originally generated in a format other than PNG it cannot be converted.
-   *
-   * @see https://www.easypost.com/docs/api/node#convert-the-label-format-of-a-shipment
-   */
-  public convertLabelFormat(format: LabelFormat): Promise<Shipment>;
-
-  /**
-   * You can update the Rates of a Shipment at any time.
-   * This operation respects the carrier_accounts attribute.
-   *
-   * @see https://www.easypost.com/docs/api/node#regenerate-rates-for-a-shipment
-   */
-  public regenerateRates(): Promise<Shipment>;
-
-  /**
-   * Insuring your Shipment is as simple as sending us the value of the contents.
-   * We charge 1% of the value, with a $1 minimum, and handle all the claims.
-   * All our claims are paid out within 10 days.
-   *
-   * To buy insurance, first purchase the Shipment, then make the insurance call before the package begins being handled by the carrier.
-   *
-   * @see https://www.easypost.com/docs/api/node#insure-a-shipment
-   *
-   * @param amount
-   */
-  public insure(amount: number): Promise<Shipment>;
-}
-
-export declare interface ICarrierAccountCreateParameters
-  extends Omit<ICarrierAccount, ParametersToOmitOnCreate> {}
-export declare class CarrierAccount implements ICarrierAccount {
-  public constructor(input: DeepPartial<ICarrierAccountCreateParameters>);
-
-  type: string;
-  fields: ICarrierAccountFields;
-  clone: boolean;
-  description: string;
-  reference: string;
-  readable: string;
-  credentials: ICarrierAccountCredentials;
-  test_credentials: ICarrierAccountCredentials;
-  id: string;
-  mode: 'test' | 'production';
-  object: 'CarrierAccount';
-  created_at: string;
-  updated_at: string;
-
-  /**
-   * CarrierAccount objects may be managed through the EasyPost API using the Production mode API key only.
-   * Multiple accounts can be added for a single carrier.
-   *
-   * The CarrierType of the preferred CarrierAccount should be consulted before attempting to create a new CarrierAccount, as it will inform you of the field names expected by a certain carrier.
-   *
-   * @see https://www.easypost.com/docs/api/node#create-a-carrier-account
-   * @see https://www.easypost.com/docs/api/node#update-a-carrieraccount
-   * @requires production API Key.
-   */
-  public save(): Promise<CarrierAccount>;
-
-  /**
-   * Retrieve an unpaginated list of all CarrierAccounts available to the authenticated account.
-   * Only Production API keys may be used to retrieve this list, as there is no test mode equivalent.
-   *
-   * @see https://www.easypost.com/docs/api/node#list-all-carrier-accounts
-   * @requires production API Key.
-   */
-  static all(): Promise<CarrierAccount[]>;
-
-  /**
-   * Retrieve a CarrierAccount by either its id or reference.
-   * However it is recommended to use EasyPost's provided identifiers because we do not enforce a unique reference.
-   *
-   * @param carrierAccountId Unique, begins with "ca_"
-   *
-   * @see https://www.easypost.com/docs/api/node#retrieve-a-carrieraccount
-   * @requires production API Key.
-   */
-  static retrieve(carrierAccountId: string): Promise<CarrierAccount>;
-
-  /**
-   * CarrierAccount objects may be removed from your account when they become out of date or no longer useful.
-   *
-   * @see https://www.easypost.com/docs/api/node#delete-a-carrier-account
-   * @requires production API Key.
-   */
-  public delete(): Promise<{}>;
-}
-
 export declare interface IOrderCreateParameters {
   reference?: string;
   to_address: IAddress | string;
@@ -1838,235 +389,6 @@ export declare class Order implements IOrder {
   public buy(carrier: Carrier, service: ServiceLevel): Promise<Order>;
 
   public getRates(): Promise<IRate[]>;
-}
-
-export declare interface IInsuranceCreateParameters {
-  reference?: string;
-  to_address: IAddress | string;
-  from_address: IAddress | string;
-
-  /**
-   * The carrier associated with the tracking_code you provided.
-   * The carrier will get auto-detected if none is provided
-   */
-  carrier?: Carrier;
-
-  /**
-   * The tracking code associated with the non-EasyPost-purchased package you'd like to insure.
-   */
-  tracking_code: string;
-
-  /**
-   * The USD value of contents you would like to insure.
-   * Currently the maximum is $5000
-   *
-   * @example 100.00
-   */
-  amount: string;
-}
-
-export declare class Insurance implements IInsurance {
-  public constructor(input: IInsuranceCreateParameters);
-
-  reference: string;
-  amount: string;
-  provider: string;
-  provider_id: string;
-  shipment_id: string;
-  tracking_code: string;
-  status: TInsuranceStatus;
-  tracker: ITracker;
-  to_address: IAddress;
-  from_address: IAddress;
-  fee: IFee;
-  messages: string[];
-  id: string;
-  mode: 'test' | 'production';
-  object: 'Insurance';
-  created_at: string;
-  updated_at: string;
-
-  /**
-   * An Insurance created via this endpoint must belong to a shipment purchased outside of EasyPost.
-   * Insurance for Shipments created within EasyPost must be created via the Shipment Buy or Insure endpoints.
-   * When creating Insurance for a non-EasyPost shipment, you must provide to_address, from_address, tracking_code, and amount information.
-   * Optionally, you can provide the carrier parameter, which will help EasyPost identify the carrier the package was shipped with.
-   * If no carrier is provided, EasyPost will attempt to determine the carrier based on the tracking_code provided.
-   * Providing a carrier parameter is recommended, since some tracking_codes are ambiguous and may match with more than one carrier.
-   * In addition, not having to auto-match the carrier will significantly speed up the response time.
-   *
-   * @see https://www.easypost.com/docs/api/node#create-an-insurance
-   */
-  public save(): Promise<Insurance>;
-
-  /**
-   * The Insurance List is a paginated list of all Insurance objects associated with the given API Key.
-   * It accepts a variety of parameters which can be used to modify the scope.
-   * The has_more attribute indicates whether or not additional pages can be requested.
-   * The recommended way of paginating is to use either the before_id or after_id parameter to specify where the next page begins.
-   *
-   * @see https://www.easypost.com/docs/api/node#retrieve-a-list-of-insurances
-   */
-  static all(
-    params?: IAllMethodParameters,
-  ): Promise<{ insurances: Insurance[]; has_more: boolean }>;
-
-  /**
-   * Retrieve an Insurance by id.
-   *
-   * @param insuranceId Unique, starts with "ins_"
-   *
-   * @see https://www.easypost.com/docs/api/node#retrieve-an-insurance
-   */
-  static retrieve(insuranceId: string): Promise<Insurance>;
-}
-
-export declare interface IPickupCreateParameters {
-  address: Address | string;
-
-  /**
-   * if no batch
-   */
-  shipment?: Shipment | string;
-
-  /**
-   * if no shipment
-   */
-  batch?: Batch | string;
-
-  carrier_accounts?: CarrierAccount[];
-  instructions?: string;
-  reference?: string;
-  is_account_address?: boolean;
-  min_datetime: string;
-  max_datetime: string;
-}
-
-export declare class Pickup implements IPickup {
-  public constructor(input: IPickupCreateParameters);
-
-  reference?: string;
-  status: 'unknown' | 'scheduled' | 'canceled';
-  min_datetime: string;
-  max_datetime: string;
-  is_account_address: boolean;
-  instructions: string;
-  messages: IMessage[];
-  confirmation: string;
-  shipment: IShipment;
-  address: IAddress;
-  carrier_accounts: ICarrierAccount[];
-  pickup_rates: IPickupRate[];
-  id: string;
-  mode: 'test' | 'production';
-  object: 'Pickup';
-  created_at: string;
-  updated_at: string;
-
-  /**
-   * Creating a Pickup will automatically fetch rates for the given time frame and location.
-   *
-   * Pickups work with existing shipments or a batch and either a fully-specified Address object or id.
-   * The examples below assume that a shipment and address have both already been created.
-   *
-   * @see https://www.easypost.com/docs/api/node#create-a-pickup
-   */
-  public save(): Promise<Pickup>;
-
-  /**
-   * A Pickup object can be retrieved by either an id or reference.
-   * However it is recommended to use EasyPost's provided identifiers because uniqueness on reference is not enforced.
-   *
-   * @param pickupId Unique, starts with "pickup_"
-   *
-   * @see https://www.easypost.com/docs/api/node#retrieve-a-pickup
-   */
-  static retrieve(pickupId: string): Promise<Pickup>;
-
-  /**
-   * To purchase a Pickup a PickupRate must be specified by its carrier and service name, instead of its id.
-   * The client libraries will handle this automatically if a PickupRate is provided.
-   *
-   * @see https://www.easypost.com/docs/api/node#buy-a-pickup
-   */
-  public buy(carrier: Carrier, service: ServiceLevel): Promise<Pickup>;
-
-  /**
-   * You may cancel a Pickup anytime before it has been completed.
-   * It requires no additional parameters other than the id or reference.
-   * The status will change to "canceled" on success.
-   *
-   * @see https://www.easypost.com/docs/api/node#cancel-a-pickup
-   */
-  public cancel(): Promise<Pickup>;
-}
-
-export declare interface IBatchCreateParameters {
-  shipments?: Array<Shipment | string>;
-}
-
-export declare class Batch implements IBatch {
-  public constructor(input: IBatchCreateParameters);
-
-  reference?: string;
-  state: TBatchState;
-  num_shipments: number;
-  shipments: IBatchShipment[];
-  status: TBatchStatuses;
-  label_url: string;
-  scan_form: IScanForm;
-  pickup: IPickup;
-  id: string;
-  mode: 'test' | 'production';
-  object: 'Batch';
-  created_at: string;
-  updated_at: string;
-
-  /**
-   * A Batch can be created with or without Shipments.
-   * When created with Shipments the initial state will be creating.
-   * Once the state changes to created a webhook Event will be sent.
-   * When created with no Shipments the initial state will be created and webhook will be sent.
-   *
-   * @see https://www.easypost.com/docs/api/node#create-a-batch
-   */
-  public save(): Promise<Batch>;
-
-  static retrieve(batchId: string): Promise<Batch>;
-
-  /**
-   * Shipments can be added to a Batch throughout its life cycle.
-   * Just remember that the state change of a Batch is asynchronous and will fire a webhook Event when the state change is completed.
-   *
-   * @see https://www.easypost.com/docs/api/node#add-shipments-to-a-batch
-   */
-  public addShipments(shipments: IBatchCreateParameters['shipments']): Promise<Batch>;
-
-  /**
-   * There could be times when a Shipment needs to be removed from the Batch during its life cycle.
-   * Removing a Shipment does not remove it from the consolidated label or ScanForm.
-   *
-   * @see https://www.easypost.com/docs/api/node#remove-shipments-from-a-batch
-   */
-  public removeShipments(shipments: IBatchCreateParameters['shipments']): Promise<Batch>;
-
-  /**
-   * One of the advantages of processing Shipments in batches is the ability to consolidate the PostageLabel into one file.
-   * This can only be done once for each batch and all Shipments must have a status of postage_purchased.
-   *
-   * Available label formats are 'pdf', 'zpl' or 'epl2' format.
-   * Like converting a PostageLabel format, if this process will change the format of the labels they must have been created as PNGs.
-   *
-   * @see https://www.easypost.com/docs/api/node#batch-labels
-   */
-  public generateLabel(labelFormat: LabelFormat): Promise<Batch>;
-
-  /**
-   * See [Scan Form](https://www.easypost.com/docs/api/node#scan-form) rules and [Object Definition](https://www.easypost.com/docs/api/node#scan-form-object).
-   *
-   * @see https://www.easypost.com/docs/api/node#manifesting-scan-form
-   */
-  public createScanForm(): Promise<Batch>;
 }
 
 export declare interface IWebhookCreateParameters {
